os2: fix searching in VPATH failure
When compiling gcc 4.9.2, compilation fails with complaining
'c/c-decl.c' not found in 'host-i686-pc-os2-emx/gcc'. This is because
the result of searching in VPATH is overridden by an original path.

That is, c/c-decl.c is found in ../.././gcc in VPATH. The result of
searching in VPATH is ../.././gcc/c-cdecl.c. However, it is overridden
by c/c-decl.c which is the path before searching in VPATH.

This is the regression of commit
f7e149812769a349ad18be2f2d0ecb52bba49b80.

    modified:   remake.c